+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 - `deleteSolidDataset` and `deleteFile`: Add functions to the `resource` module
   to delete resources, following the interface of `@inrupt/solid-client`.
+- `getAccessRequest`: a function exported by the `./manage` module to
+  get the Access Request from the Access Request URL.
 
 ## [2.3.2](https://github.com/inrupt/solid-client-access-grants-js/releases/tag/v2.3.2) - 2023-06-05